pyproject.toml
[tool.poetry]
name = "frikanalen-django-api-client"
version = "1.1.0"
description = "A client library for accessing Frikanalen Django API"
authors = []
readme = "README.md"
packages = [
    {include = "frikanalen_django_api_client"},
]
include = ["CHANGELOG.md", "frikanalen_django_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
